ci(docker-build-and-publish): allow transferring builds for pull requests
Using a combination of GitHub artifacts and a local registry we can
transfer built images between all jobs in the workflow.
This allows building the image in pull requests to verify that the build
succeeds before merging the pull request.
Sadly this is the only way to do this as GitHub does not allow the
images to be pushed to the fork maintainers GHCR which would make this
much simpler.